+++ src/Resources/contao/dca/tl_page.php
@@ -1,0 +1,8 @@
+<?php
+$GLOBALS['TL_DCA']['tl_page']['list']['global_operations']['counter_pages'] = array
+(
+	'label'               => &$GLOBALS['TL_LANG']['tl_page']['counter_pages'],
+	'href'                => 'key=counter',
+	'icon'                => 'bundles/contaocounter/images/counter.png',
+	'attributes'          => 'onclick="Backend.getScrollOffset();"'
+);