+++ src/Resources/contao/languages/de/tl_page.php
@@ -1,0 +1,3 @@
+<?php
+
+$GLOBALS['TL_LANG']['tl_page']['counter_pages'] = array('Statistik', 'Seitenstatistik anzeigen');