+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,12 @@
+<?php
+
+/**
+ * Legende
+ */
+$GLOBALS['TL_LANG']['tl_settings']['counter_legend']       = 'Zähler';
+
+/**
+ * Felder
+ */
+$GLOBALS['TL_LANG']['tl_settings']['counter_topx_news'] = array('Anzahl Nachrichten', 'Anzahl der Nachrichten in der Top-Liste der Statistik');
+$GLOBALS['TL_LANG']['tl_settings']['counter_topx_pages'] = array('Anzahl Seiten', 'Anzahl der Seiten in der Top-Liste der Statistik');